COMP: Fixed the warning messages on VS14.
- Fixed the warning messages of data-conversion(C4267) on VS14 with
  "static_cast" and itk::Math::CastWithRangeCheck<>()".

- Almost warning messages are related to "conversion from size_t to ***
  data types".
  Please check the following list of warning messages:

- Removed the macro of "#pragma warning ( disable : 4267 )" from the file
  of "itkWin32Header.h"

  Warning C4267	conversion from 'size_t' to 'unsigned int'
  Warning C4267	conversion from 'size_t' to 'int'
  Warning C4267	conversion from 'size_t' to 'unsigned long',
  Warning C4267	conversion from 'size_t' to 'const unsigned long',
  Warning C4267	conversion from 'size_t' to 'itk::SizeValueType',
  Warning C4267	conversion from 'size_t' to 'const itk::SizeValueType'
  Warning C4267	conversion from 'size_t' to
  'itk::CompositeTransform<TParametersValueType,3>::NumberOfParametersType'
  Warning C4267	conversion from 'size_t' to
   'itk::ImageIORegion::SizeValueType'
  Warning C4267	conversion from 'size_t' to
   'itk::RingBuffer<itk::DataObject>::SizeValueType
  Warning C4267	conversion from 'size_t' to
   'itk::MultiTransform<TParametersValueType,3,3>::NumberOfParametersType'
  Warning C4267	conversion from 'size_t' to
   'itk::ImageIOBase::SizeValueType'
  Warning C4267	conversion from 'size_t' to
   'itk::ParticleSwarmOptimizerBase::NumberOfParticlesType'
  Warning C4267	conversion from 'size_t' to
   'itk::Array<int>::SizeValueType'
  Warning C4267	conversion from 'size_t' to
   'itk::VideoIOBase::FrameOffsetType'
  Warning C4267	conversion from 'size_t' to
  'itk::ByteSwapper<ValueType>::BufferSizeType'
  Warning C4267	conversion from 'size_t' to
   'itk::MultiTransform<TParametersValueType,3,3>::NumberOfParametersType'
  Warning C4267	conversion from 'size_t' to
  'itk::Statistics::ListSample<TargetVectorType>::InstanceIdentifier',
  Warning C4267	conversion from 'size_t' to
   'itk::ByteSwapper<ValueType>::BufferSizeType'
  Warning C4267	conversion from 'size_t' to
   'itk::Array<unsigned int>::SizeValueType'
  .....

+++ Examples/IO/XML/itkParticleSwarmOptimizerSAXWriter.cxx
@@ -118,9 +118,11 @@ int ParticleSwarmOptimizerSAXWriter::WriteFile()
     ofs << "  <ParametersConvergenceTolerance>";
 
     Array<double> ptols = this->m_InputObject->GetParametersConvergenceTolerance();
-    for ( size_t i = 0; i < ptols.GetSize(); i++ )
+    // Note: The data-cast to unsigned int is required
+    //       because itk::Array only supports 'unsigned int' number of elements.
+    for ( unsigned int i = 0; i < ptols.GetSize(); i++ )
       {
-      ofs << " " << ptols[i];
+      ofs << " " << ptols[ i ];
       }
 
     ofs << "</ParametersConvergenceTolerance>";
